+++ src/Clash/Cores/Arp/ArpManager.hs
@@ -9,6 +9,7 @@ Description : Provides an ARP manager which handles ARP lookups from client circ
 module Clash.Cores.Arp.ArpManager where
 
 import Clash.Prelude
+import Clash.Signal.Extra
 
 import Protocols
 import Protocols.Df qualified as Df
@@ -17,24 +18,6 @@ import Clash.Cores.Arp.ArpTypes
 import Clash.Cores.Ethernet.EthernetTypes
 import Clash.Cores.IP.IPv4Types
 
-
--- | This register is @True@ exactly every second if @DomainPeriod dom@ divides @10^12@.
---   If not, the accuracy depends on the clock frequency, because we round this division
---   down. In that case, the higher the clock frequency, the more accurate it is.
---   Does not support clock frequencies lower than 2 Hz.
-secondTimer
-  :: forall (dom :: Domain)
-   . HiddenClockResetEnable dom
-  => KnownNat (DomainPeriod dom)
-  => 1 <= DomainPeriod dom
-  => DomainPeriod dom <= 5 * 10^11
-  => Signal dom Bool
-secondTimer = case compareSNat d1 (SNat @(10^12 `Div` DomainPeriod dom)) of
-  SNatLE -> isRising 0 $ msb <$> counter
-    where
-      counter :: Signal dom (Index (10^12 `Div` DomainPeriod dom))
-      counter = register maxBound (satPred SatWrap <$> counter)
-  SNatGT -> errorX "secondTimer: Absurd, Report this to the Clash compiler team: https://github.com/clash-lang/clash-compiler/issues"
 
 -- | State of the ARP manager.
 data ArpManagerState maxWaitSeconds
